# Optimize `Array#concat(Indexable)`