@HertzDevil 2023-04-04
There is a compilation error that only affects [Crystal 1.0.0](https://play.crystal-lang.org/#/r/etnb)? Same piece of code works on [Crystal 1.1.1](https://play.crystal-lang.org/#/r/etna)